ai-model/requirements.txt
tensorflow==2.20.0
Pillow>=9.0.0
numpy>=1.21.0